Use OAuth for authentication
This uses the emacs-oauth package from Peter Sanford to authenticate
against Twitter. This is needed since Twitter disabled basic
authentication recently. OAuth requires a consumer key and consumer
secret in order to get an access token. However, the application is
meant to keep the key and secret private but of course there is no way
to hide this in an open source application. I've just left these two
parameters as custom variables in the hope that the user will find some
way to get their own values.

The access token is stored in the file ~/.twitel-access-token so that
the authentication process doesn't have to happen every time twitel is
used.

The username and password variables are removed since Twitel no longer
needs to send them to authenticate.